Add tests for main function of moco-controller

+++ e2e/agent_test.go
@@ -25,7 +25,9 @@ func testAgent() {
 	})
 
 	AfterEach(func() {
-		portForwardCmd.Process.Kill()
+		if portForwardCmd != nil {
+			portForwardCmd.Process.Kill()
+		}
 	})
 
 	It("should run health probe server", func() {
